pyproject.toml
[project]
name = "bot-core-lib"
version = "0.1.0"
description = "Shared core library for bots"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
